Support unity build for unittests
