+++ unittests/special_accounts_tests.cpp
@@ -31,12 +31,12 @@ BOOST_FIXTURE_TEST_CASE(accounts_exists, tester)
 
       auto nobody = chain1_db.find<account_object, by_name>(config::null_account_name);
       BOOST_CHECK(nobody != nullptr);
-      const auto& nobody_active_authority = chain1_db.get<permission_object, by_owner>(boost::make_tuple(config::null_account_name, config::active_name));
+      const auto& nobody_active_authority = chain1_db.get<permission_object, eosio::chain::by_owner>(boost::make_tuple(config::null_account_name, config::active_name));
       BOOST_CHECK_EQUAL(nobody_active_authority.auth.threshold, 1u);
       BOOST_CHECK_EQUAL(nobody_active_authority.auth.accounts.size(), 0u);
       BOOST_CHECK_EQUAL(nobody_active_authority.auth.keys.size(), 0u);
 
-      const auto& nobody_owner_authority = chain1_db.get<permission_object, by_owner>(boost::make_tuple(config::null_account_name, config::owner_name));
+      const auto& nobody_owner_authority = chain1_db.get<permission_object, eosio::chain::by_owner>(boost::make_tuple(config::null_account_name, config::owner_name));
       BOOST_CHECK_EQUAL(nobody_owner_authority.auth.threshold, 1u);
       BOOST_CHECK_EQUAL(nobody_owner_authority.auth.accounts.size(), 0u);
       BOOST_CHECK_EQUAL(nobody_owner_authority.auth.keys.size(), 0u);
