+++ unittests/test_cfd_transaction.cpp
@@ -1,29 +1,7 @@
 #include "test_cfd_transaction.hpp"
 #include <contracts.hpp>
 
-struct dummy_action {
-   static eosio::chain::name get_name() { return N(dummyaction); }
-   static eosio::chain::name get_account() { return N(testapi); }
 
-   char     a; // 1
-   uint64_t b; // 8
-   int32_t  c; // 4
-};
-
-struct cf_action {
-   static eosio::chain::name get_name() { return N(cfaction); }
-   static eosio::chain::name get_account() { return N(testapi); }
-
-   uint32_t payload = 100;
-   uint32_t cfd_idx = 0; // context free data index
-};
-
-FC_REFLECT(dummy_action, (a)(b)(c))
-FC_REFLECT(cf_action, (payload)(cfd_idx))
-
-#define DUMMY_ACTION_DEFAULT_A 0x45
-#define DUMMY_ACTION_DEFAULT_B 0xab11cd1244556677
-#define DUMMY_ACTION_DEFAULT_C 0x7451ae12
 
 std::vector<eosio::chain::signed_block_ptr> deploy_test_api(eosio::testing::tester& chain) {
    std::vector<eosio::chain::signed_block_ptr> result;
