+++ unittests/wasm_tests.cpp
@@ -67,16 +67,14 @@ struct provereset {
    }
 };
 
-namespace {
+FC_REFLECT_EMPTY(provereset);
+
+BOOST_AUTO_TEST_SUITE(wasm_tests)
+
 #warning Change this back to using TESTER
 struct old_wasm_tester : tester {
    old_wasm_tester() : tester{setup_policy::old_wasm_parser} {}
 };
-}
-
-FC_REFLECT_EMPTY(provereset);
-
-BOOST_AUTO_TEST_SUITE(wasm_tests)
 
 /**
  * Prove that action reading and assertions are working
@@ -180,14 +178,14 @@ BOOST_FIXTURE_TEST_CASE( abi_from_variant, TESTER ) try {
    set_abi(N(asserter), contracts::asserter_abi().data());
    produce_blocks(1);
 
-   auto resolver = [&,this]( const account_name& name ) -> optional<abi_serializer> {
+   auto resolver = [&,this]( const account_name& name ) -> fc::optional<abi_serializer> {
       try {
          const auto& accnt  = this->control->db().get<account_object,by_name>( name );
          abi_def abi;
          if (abi_serializer::to_abi(accnt.abi, abi)) {
             return abi_serializer(abi, abi_serializer::create_yield_function( abi_serializer_max_time ));
          }
-         return optional<abi_serializer>();
+         return fc::optional<abi_serializer>();
       } FC_RETHROW_EXCEPTIONS(error, "Failed to find or parse ABI for ${name}", ("name", name))
    };
